api/backend/requirements.txt
Django==5.0.4
django_environ==0.11.2
djangorestframework==3.15.1
environ==1.0
google_cloud_aiplatform==1.48.0
pandas==2.2.2
Requests==2.31.0
vertexai==1.48.0
